pyproject.toml
[project]
name = "mssql-coding-case"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.1",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "pyjwt>=2.10.1",
    "pymssql>=2.3.4",
    "python-dotenv>=1.1.0",
]
